Create README.md
This study uses a Long Short-Term Memory (LSTM) model to examine water usage cost data for the New York City boroughs. Data on past water use costs is gathered to guarantee uniformity and quality, and preprocessed. Future water consumption expenses are forecasted using a model that is built using a specific LSTM architecture that has been constructed and trained. The findings of this research help to increase understanding of water consumption cost trends in the boroughs of New York, which supports the development of policies, long-term planning, and evidence-based decision-making for sustainable water management. The objective is to address the issues related to water management in the twenty-first century using advanced digital technology and innovative approaches.